+++ CHANGELOG.md
@@ -104,5 +104,5 @@ and fix tailwindcss issues.
 ### Added
 
 - `Input` component refactored to use hide mask when input is focused.
-
+- `Input` component refactored to use auto scroll to the end of the input when the input is focused.